betting_backend/requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# API clients
requests==2.31.0
nba-api==1.4.1

# Data processing
numpy==1.26.2
pandas==2.1.4

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# Utilities
python-dateutil==2.8.2

# Development and testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
